chore: restructure project and update package declarations
