+++ main/java/javaroke/models/NodeSong.java
@@ -1,6 +1,6 @@
-package models;
+package javaroke.models;
 
-import utils.Validator;
+import javaroke.utils.Validator;
 
 public class NodeSong {
     private String songId;
